% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{a1.}
\alias{a1.}
\title{Shortcut for apply(x, 1, function(.) expr)}
\usage{
a1.(x, expr)
}
\arguments{
\item{x}{something you can call apply on, i.e. a matrix, a data.frame a data.table...}

\item{expr}{expr any R expression that is evaluated on every row, ., of x}
}
\value{
eval(substitute(apply(xx, 1, function(.) expr)), list(xx=x))
}
\description{
Shortcut for apply(x, 1, function(.) expr)
}
