% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{a2.}
\alias{a2.}
\title{Shortcut for apply(x, 2, function(.) expr)}
\usage{
a2.(x, expr)
}
\arguments{
\item{x}{something you can call apply on, i.e. a matrix, a data.frame a data.table...}

\item{expr}{expr any R expression that is evaluated on every row, ., of x}
}
\value{
eval(substitute(apply(xx, 2, function(.) expr)), list(xx=x))
}
\description{
Shortcut for apply(x, 2, function(.) expr)
}
