% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyseBenchData}
\alias{analyseBenchData}
\title{Apply the function what on all lines specified by filter in benchdata}
\usage{
analyseBenchData(benchdata, filter, what, ...)
}
\arguments{
\item{benchdata}{a data.table}

\item{filter}{a list of elements with modes corresponding to the first key-columns of
benchdata}

\item{what}{a function(row) returning a list of statistics for a row of benchdata}
}
\description{
Apply the function what on all lines specified by filter in benchdata
}

