% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{collectBenchRes}
\alias{collectBenchRes}
\title{Collect benchmark results from .RData files}
\usage{
collectBenchRes(fs, benchdata = NULL, table.file = NULL, ids = NULL,
  res.ids.only = T, dir.res = ".", write.file = F, verbose = F)
}
\arguments{
\item{fs}{a character vector containing function names used for executing 
previous benchmarks}

\item{benchdata}{a data.table where the collected results will be stored as
columns, defaults to NULL in which case table.file should be specified.}

\item{table.file}{a file containing a data.table object where the 
results will be added to, defaults to NULL in which case benchdata should be
specified.}

\item{ids}{NULL or vector of integers or a list of integer vectors of length
equal to length(fs) containing id's of lines for which results are to be 
collected.}

\item{res.ids.only}{logical indicating whether to restrict the result to the 
lines indicated by ids. Default is TRUE.}

\item{dir.res}{character indicating the directory where result-files are 
stored, default is '.'.}

\item{write.file}{a logical indicating whether to write the resulting 
data.table to the file table.file, defaults to FALSE.}

\item{verbose}{logical indicating whether some file-loading messages should be output}
}
\value{
the resulting data.table object with added (or updated) columns for 
each element of fs.
}
\seealso{
\code{\link{genBenchJobs}}
}

