% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{doBenchJob}
\alias{doBenchJob}
\title{Extract command-line arguments specifying a call to benchmark.}
\usage{
doBenchJob(fname, f, table.file, table.name, ids, ...)
}
\arguments{
\item{fname, f, table.file, table.name, ids, ...}{internal usage (see code).}
}
\description{
Extract command-line arguments specifying a call to benchmark.
}
\note{
This is an internal function
}
