% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{getLineAsList}
\alias{getLineAsList}
\title{Get a list representing a line of a data.table}
\usage{
getLineAsList(benchdata, i)
}
\arguments{
\item{benchdata}{a data.table with a column called 'id' as key}

\item{i}{a single integer specifying the id of a line}
}
\value{
a named list of objects found in the columns of benchdata at the 
  specified line.
}
\description{
Get a list representing a line of a data.table
}
