% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{l.}
\alias{l.}
\title{Shortcut for lapply(l, function(.) expr)}
\usage{
l.(l, expr)
}
\arguments{
\item{l}{a list}

\item{expr}{an expression that is applied to every element, ., of l, e.g. 5*.}
}
\value{
eval(substitute(lapply(ll, function(.) expr)), list(ll=l))
}
\description{
Shortcut for lapply(l, function(.) expr)
}

