% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{listBenchAnalysis}
\alias{listBenchAnalysis}
\title{Create a list of matrices with results from analyseBenchData as columns}
\usage{
listBenchAnalysis(benchdata, filter, what, levels)
}
\arguments{
\item{benchdata}{a data.table with a key}

\item{filter}{a list of elements with modes corresponding to the first key-columns of
benchdata}

\item{what}{a function(row) returning a list of statistics for a row of benchdata}

\item{levels}{a vector with elements corresponding to the columns of the matrices. The
mode of this vector should correspond to the mode of a key-column in benchdata}
}
\description{
Create a list of matrices with results from analyseBenchData as columns
}
