% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lll.}
\alias{lll.}
\title{Construct a data.frame from its arguments and evaluate expression on every row}
\usage{
lll.(l1 = NULL, l2 = NULL, expr, ...)
}
\arguments{
\item{l1, }{l2, ... two or more vector or list objects. A data.frame is constructed from these
by data.frame(l1, l2, ...).}

\item{expr}{any R expression that is evaluated on every row, ., of the data.frame.}
}
\value{
eval(substitute(alply(dt, 1, function(.) expr)))
}
\description{
Construct a data.frame from its arguments and evaluate expression on every row
}
