% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{s.}
\alias{s.}
\title{Shortcut for sapply(l, function(.) expr)}
\usage{
s.(l, expr)
}
\arguments{
\item{l}{a list}

\item{expr}{an expression that is evaluated to every element, ., of l, e.g. 5*.}
}
\value{
eval(substitute(sapply(ll, function(.) expr)), list(ll=l))
}
\description{
Shortcut for sapply(l, function(.) expr)
}
