% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{transposeAnForBoxPlot}
\alias{transposeAnForBoxPlot}
\title{transpose an into a list of S matrices of length(levels) columns}
\usage{
transposeAnForBoxPlot(an, levels = names(an),
  namesBoxes = names(an[[1]][[1]]))
}
\arguments{
\item{an}{a named list of analysis returned by analyseBendData}

\item{levels}{character or numeric corresponding to box-plot-groups to be plotted along the x-axis}

\item{namesBoxes}{character corresponding to different boxes in each box-plot-group.}
}
\value{
a named list of matrices suitable for bplListBenchAnalysis and vioplListBenchAnalysis
}
\description{
transpose an into a list of S matrices of length(levels) columns
}
